Tools: Add workaround for lupdate missing source
lupdate requires at least one C++ source file to be present in order
to process the project file. In cases where a workbench is Python-only,
this resulted in incorrect or missing extraction of strings in ui files.
Those files should have been processed by lupdate, but were not.
pyludate was not configured to process them. To work around this, the
updatets.py script is modified to create a dummy *.cpp file before
running qmake.