Github: rename Workbench tags so they present alphabetically in GH UI
When choosing a specific workbench tag in the GH the current dropdown entries are scattered. By prepending `WB` to them and sorting them alphabetically will make it cleaner to select them. Due to this, the tags have been renamed.

This PR follows up the change of the tags to the GH auto-labeler action. Some extra tags were added as well like :octocat: for GH files and `Addon Manager` etc...

[skip ci]