Fix build of "tests" on macOS
Building the C++ unit tests failed for
 tests/src/App/Metadata.cpp
 tests/src/Base/Reader.cpp
because some XML headers were not found. This was because the XercesC
include dir was not in the search list for headers.
I cannot say why only macOS seems to be affected, but it's not the first
time, see e.g https://github.com/FreeCAD/FreeCAD/commit/6fd90bcd26ba086b6052b2920c1c926018ec782d