Finish renaming desktop filename to org.freecad.FreeCAD
The QGuiApplication::setDesktopFileName() call was changed in [1] but
the desktop file and other files that reference it were not.

As shown by Looking Glass [2] on X11 updating these files is still
not enough for GNOME to associate the process with the desktop file,
a StartupWMClass entry is added to the desktop file to fix this.

[1]
2202254d7a ("Migrate domain name from freecadweb to freecad (#9352)", 2023-04-24)

[2]
https://blogs.gnome.org/mclasen/2016/11/15/help-my-app-icon-is-missing/