Spreadsheet: fix build on macOS
The recent migration from boost::bind to std::bind [1] broke the build
on Mac (Apple clang version 13.0.0 (clang-1300.0.29.30) on macOS-11.7.8).

In all other cases the `boost::bind` was replaced by `std::bind` (among
with the place holders) but in these two spread sheet files
  src/Mod/Spreadsheet/Gui/SheetModel.cpp
  src/Mod/Spreadsheet/Gui/SpreadsheetView.cpp
the original code only referenced `bind`, which used to get resolved to
`boost::bind` but now raises this error:
> /Users/jonas/src/FreeCAD/FreeCAD-git/src/Mod/Spreadsheet/Gui/SheetModel.cpp:50:36: error: use of undeclared identifier 'bind'; did you mean 'boost::bind'?
>         sheet->cellUpdated.connect(bind(&SheetModel::cellUpdated, this, sp::_1));
>                                    ^~~~
>                                    boost::bind

This commit changes this to `std::bind` expicitly, just like it's done in
the other files. This works on my system/compiler.

[1]: https://github.com/FreeCAD/FreeCAD/commit/232abc60fe2cd5e86f5244896255ab0cb5887d69